refactor HttpRequestParser
